Abstract core module maintainer api to AbstractCoreMaintainerService
