+++ maintainer-client/src/main/java/com/alibaba/nacos/maintainer/client/config/ConfigMaintainerService.java
@@ -17,6 +17,7 @@
 package com.alibaba.nacos.maintainer.client.config;
 
 import com.alibaba.nacos.api.exception.NacosException;
+import com.alibaba.nacos.maintainer.client.core.CoreMaintainerService;
 import com.alibaba.nacos.maintainer.client.model.config.Capacity;
 import com.alibaba.nacos.maintainer.client.model.config.ConfigAdvanceInfo;
 import com.alibaba.nacos.maintainer.client.model.config.ConfigAllInfo;
@@ -28,12 +29,7 @@ import com.alibaba.nacos.maintainer.client.model.config.GroupkeyListenserStatus;
 import com.alibaba.nacos.maintainer.client.model.config.Page;
 import com.alibaba.nacos.maintainer.client.model.config.SameConfigPolicy;
 import com.alibaba.nacos.maintainer.client.model.config.SameNamespaceCloneConfigBean;
-import com.alibaba.nacos.maintainer.client.model.core.Connection;
-import com.alibaba.nacos.maintainer.client.model.core.IdGeneratorVO;
-import com.alibaba.nacos.maintainer.client.model.core.Member;
-import com.alibaba.nacos.maintainer.client.model.core.ServerLoaderMetrics;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
@@ -42,7 +38,7 @@ import java.util.Map;
  *
  * @author Nacos
  */
-public interface ConfigMaintainerService {
+public interface ConfigMaintainerService extends CoreMaintainerService {
     
     /**
      * Get configuration information by dataId and groupName.
@@ -398,124 +394,4 @@ public interface ConfigMaintainerService {
      */
     Map<String, Object> getClusterMetrics(String ip, String dataId, String groupName, String namespaceId)
             throws NacosException;
-    
-    // ------------------------- Core Operations -------------------------
-    
-    /**
-     * Execute a Raft operation with the specified command, value, and group ID.
-     *
-     * @param command the command to execute.
-     * @param value   the value associated with the command.
-     * @param groupId the group ID for the operation.
-     * @return the result of the Raft operation.
-     * @throws NacosException if the operation fails.
-     */
-    String raftOps(String command, String value, String groupId) throws NacosException;
-    
-    /**
-     * Retrieve the current health status of the ID generator.
-     *
-     * @return a list of ID generator status objects.
-     * @throws NacosException if the operation fails.
-     */
-    List<IdGeneratorVO> getIdsHealth() throws NacosException;
-    
-    /**
-     * Update the log level for a specific logger.
-     *
-     * @param logName  the name of the logger to update.
-     * @param logLevel the new log level to set.
-     * @throws NacosException if the operation fails.
-     */
-    void updateLogLevel(String logName, String logLevel) throws NacosException;
-    
-    /**
-     * Retrieve information about the current node.
-     *
-     * @return the current node's information.
-     * @throws NacosException if the operation fails.
-     */
-    Member getSelfNode() throws NacosException;
-    
-    /**
-     * List cluster nodes based on the specified address and state.
-     *
-     * @param address the address to filter nodes by.
-     * @param state   the state to filter nodes by.
-     * @return a collection of matching nodes.
-     * @throws NacosException if an error occurs during the operation.
-     */
-    Collection<Member> listClusterNodes(String address, String state) throws NacosException;
-    
-    /**
-     * Retrieve the health status of the current node.
-     *
-     * @return the health status of the current node.
-     * @throws NacosException if the operation fails.
-     */
-    String getSelfNodeHealth() throws NacosException;
-    
-    /**
-     * Update the list of cluster nodes.
-     *
-     * @param nodes the list of nodes to update.
-     * @return true if the operation was successful, false otherwise.
-     * @throws NacosException if an error occurs during the operation.
-     */
-    Boolean updateClusterNodes(List<Member> nodes) throws NacosException;
-    
-    /**
-     * Update the lookup mode for the cluster.
-     *
-     * @param type the type of lookup mode to set.
-     * @return true if the operation was successful, false otherwise.
-     * @throws NacosException if an error occurs during the operation.
-     */
-    Boolean updateLookupMode(String type) throws NacosException;
-    
-    /**
-     * Retrieve the current client connections.
-     *
-     * @return a map of current client connections.
-     * @throws NacosException if the operation fails.
-     */
-    Map<String, Connection> getCurrentClients() throws NacosException;
-    
-    /**
-     * Reload the number of SDK connections on the current server.
-     *
-     * @param count           the number of connections to reload.
-     * @param redirectAddress the address to redirect connections to.
-     * @return the result of the operation.
-     * @throws NacosException if the operation fails.
-     */
-    String reloadConnectionCount(Integer count, String redirectAddress) throws NacosException;
-    
-    /**
-     * Smartly reload the cluster based on the specified loader factor.
-     *
-     * @param loaderFactorStr the loader factor string.
-     * @return the result of the operation.
-     * @throws NacosException if the operation fails.
-     */
-    String smartReloadCluster(String loaderFactorStr) throws NacosException;
-    
-    /**
-     * Reload a single client connection.
-     *
-     * @param connectionId    the ID of the connection to reload.
-     * @param redirectAddress the address to redirect the connection to.
-     * @return the result of the operation.
-     * @throws NacosException if the operation fails.
-     */
-    String reloadSingleClient(String connectionId, String redirectAddress) throws NacosException;
-    
-    /**
-     * Retrieve the current cluster loader metrics.
-     *
-     * @return the loader metrics for the cluster.
-     * @throws NacosException if the operation fails.
-     */
-    ServerLoaderMetrics getClusterLoaderMetrics() throws NacosException;
-    
 }
