+++ maintainer-client/src/test/java/com/alibaba/nacos/maintainer/client/naming/NacosNamingMaintainerServiceImplTest.java
@@ -24,22 +24,19 @@ import com.alibaba.nacos.api.naming.pojo.maintainer.InstanceMetadataBatchResult;
 import com.alibaba.nacos.api.naming.pojo.maintainer.MetricsInfo;
 import com.alibaba.nacos.api.naming.pojo.maintainer.ServiceDetailInfo;
 import com.alibaba.nacos.common.http.HttpRestResult;
-import com.alibaba.nacos.maintainer.client.model.core.Connection;
-import com.alibaba.nacos.maintainer.client.model.core.IdGeneratorVO;
-import com.alibaba.nacos.maintainer.client.model.core.Member;
-import com.alibaba.nacos.maintainer.client.model.core.ServerLoaderMetrics;
+import com.alibaba.nacos.maintainer.client.core.AbstractCoreMaintainerService;
 import com.alibaba.nacos.maintainer.client.remote.ClientHttpProxy;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -47,12 +44,12 @@ import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+@ExtendWith(MockitoExtension.class)
 public class NacosNamingMaintainerServiceImplTest {
     
     @Mock
@@ -62,11 +59,10 @@ public class NacosNamingMaintainerServiceImplTest {
     
     @BeforeEach
     void setUp() throws Exception {
-        MockitoAnnotations.openMocks(this);
         Properties properties = new Properties();
         properties.setProperty("serverAddr", "localhost:8848");
         nacosNamingMaintainerService = new NacosNamingMaintainerServiceImpl(properties);
-        Field clientHttpProxyField = NacosNamingMaintainerServiceImpl.class.getDeclaredField("clientHttpProxy");
+        Field clientHttpProxyField = AbstractCoreMaintainerService.class.getDeclaredField("clientHttpProxy");
         clientHttpProxyField.setAccessible(true);
         clientHttpProxyField.set(nacosNamingMaintainerService, clientHttpProxy);
     }
@@ -701,234 +697,6 @@ public class NacosNamingMaintainerServiceImplTest {
         
         // Act
         ObjectNode result = nacosNamingMaintainerService.getResponsibleServerForClient(ip, port);
-        
-        // Assert
-        assertNotNull(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testRaftOps() throws Exception {
-        // Arrange
-        String command = "testCommand";
-        String value = "testValue";
-        String groupId = "testGroup";
-        
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>("success")));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        String result = nacosNamingMaintainerService.raftOps(command, value, groupId);
-        
-        // Assert
-        assertEquals("success", result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testGetIdsHealth() throws Exception {
-        // Arrange
-        List<IdGeneratorVO> expectedList = new ArrayList<>();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedList)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        List<IdGeneratorVO> result = nacosNamingMaintainerService.getIdsHealth();
-        
-        // Assert
-        assertNotNull(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testUpdateLogLevel() throws Exception {
-        // Arrange
-        String logName = "testLog";
-        String logLevel = "INFO";
-        
-        // Act
-        nacosNamingMaintainerService.updateLogLevel(logName, logLevel);
-        
-        // Assert
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testGetSelfNode() throws Exception {
-        // Arrange
-        Member expectedMember = new Member();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedMember)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        Member result = nacosNamingMaintainerService.getSelfNode();
-        
-        // Assert
-        assertNotNull(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testListClusterNodes() throws Exception {
-        // Arrange
-        String address = "127.0.0.1:8848";
-        String state = "UP";
-        
-        Collection<Member> expectedMembers = new ArrayList<>();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedMembers)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        Collection<Member> result = nacosNamingMaintainerService.listClusterNodes(address, state);
-        
-        // Assert
-        assertNotNull(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testGetSelfNodeHealth() throws Exception {
-        // Arrange
-        String expectedHealth = "HEALTHY";
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedHealth)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        String result = nacosNamingMaintainerService.getSelfNodeHealth();
-        
-        // Assert
-        assertEquals(expectedHealth, result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testUpdateClusterNodes() throws Exception {
-        // Arrange
-        List<Member> nodes = new ArrayList<>();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(true)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        Boolean result = nacosNamingMaintainerService.updateClusterNodes(nodes);
-        
-        // Assert
-        assertTrue(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testUpdateLookupMode() throws Exception {
-        // Arrange
-        String type = "testType";
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(true)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        Boolean result = nacosNamingMaintainerService.updateLookupMode(type);
-        
-        // Assert
-        assertTrue(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testGetCurrentClients() throws Exception {
-        // Arrange
-        Map<String, Connection> expectedMap = new HashMap<>();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedMap)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        Map<String, Connection> result = nacosNamingMaintainerService.getCurrentClients();
-        
-        // Assert
-        assertNotNull(result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testReloadConnectionCount() throws Exception {
-        // Arrange
-        Integer count = 10;
-        String redirectAddress = "localhost:8848";
-        
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>("success")));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        String result = nacosNamingMaintainerService.reloadConnectionCount(count, redirectAddress);
-        
-        // Assert
-        assertEquals("success", result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testSmartReloadCluster() throws Exception {
-        // Arrange
-        String loaderFactorStr = "testFactor";
-        
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>("success")));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        String result = nacosNamingMaintainerService.smartReloadCluster(loaderFactorStr);
-        
-        // Assert
-        assertEquals("success", result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testReloadSingleClient() throws Exception {
-        // Arrange
-        String connectionId = "testConnectionId";
-        String redirectAddress = "localhost:8848";
-        
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>("success")));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        String result = nacosNamingMaintainerService.reloadSingleClient(connectionId, redirectAddress);
-        
-        // Assert
-        assertEquals("success", result);
-        verify(clientHttpProxy, times(1)).executeSyncHttpRequest(any());
-    }
-    
-    @Test
-    void testGetClusterLoaderMetrics() throws Exception {
-        // Arrange
-        ServerLoaderMetrics expectedMetrics = new ServerLoaderMetrics();
-        HttpRestResult<String> mockHttpRestResult = new HttpRestResult<>();
-        mockHttpRestResult.setData(new ObjectMapper().writeValueAsString(new Result<>(expectedMetrics)));
-        
-        when(clientHttpProxy.executeSyncHttpRequest(any())).thenReturn(mockHttpRestResult);
-        
-        // Act
-        ServerLoaderMetrics result = nacosNamingMaintainerService.getClusterLoaderMetrics();
         
         // Assert
         assertNotNull(result);
